moving outreach code that cedric pushed to outreach, polishing some things after william and harry push
