+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Outreach/OutreachClawbot.java
@@ -1,5 +1,4 @@
-package org.firstinspires.ftc.teamcode.Old.Teleop;
-
+package org.firstinspires.ftc.teamcode.Outreach;
 
 
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
@@ -85,3 +84,4 @@ public class OutreachClawbot extends LinearOpMode {
 
     }
 }
+
